proj

+++ libft/ft_strlen.c
@@ -1,8 +1,8 @@
-#include <libft.h>
+#include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	c;
+	size_t	c;
 
 	c = 0;
 	while (str[c] != 0)
